# contracts/openapi-public.yaml
openapi: 3.1.0
info:
  title: SmartDiag504 Platform API
  version: 0.4.0
  description: >-
    Contrato de las superficies públicas, de operación y administración de SmartDiag504.
    ERPNext conserva la autoridad financiera y de inventario; esta API coordina catálogo,
    reservas, chatbot, pedidos web y proyección operacional.
servers:
  - url: https://api.smartdiag504.com
security: []
...
paths:
...
  /api/v1/store/orders:
    post:
      tags: [store]
      operationId: createStoreOrderRequest
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StoreOrderCreate'}
      responses:
        '201':
          description: Solicitud de pedido creada; todavía no constituye factura ni pago.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StoreOrder'}
        '200':
          description: Se devolvió el pedido idempotente existente.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StoreOrder'}
        '409': {description: Precio, disponibilidad o producto cambió.}
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
...
  schemas:
...
    StoreOrderItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 20}
    StoreOrderCreate:
      type: object
      required: [customer_name, phone, idempotency_key, items]
      properties:
        customer_name: {type: string, minLength: 2, maxLength: 120}
        phone: {type: string, minLength: 7, maxLength: 30}
        email: {type: [string, 'null'], format: email}
        vehicle_vin: {type: [string, 'null'], minLength: 11, maxLength: 17}
        notes: {type: [string, 'null'], maxLength: 2000}
        idempotency_key: {type: string, minLength: 8, maxLength: 128}
        items:
          type: array
          minItems: 1
          maxItems: 50
          items: {$ref: '#/components/schemas/StoreOrderItemCreate'}
    StoreOrder:
      type: object
      required: [id, order_number, status, currency, subtotal, items]
      properties:
        id: {type: string, format: uuid}
        order_number: {type: string}
        status: {enum: [PENDING_CONFIRMATION, CONTACTED, CONFIRMED, SYNCED, CANCELLED]}
        currency: {type: string}
        subtotal: {type: string}
        erpnext_sales_order_id: {type: [string, 'null']}
        items: {type: array, items: {type: object}}
...